requirements.txt
django
django-tailwind
django-multiselectfield